some fixes for living behind a reverse proxy
... you cannot use the req.originalUrl when the Node app is behind a
reverse proxy. Instead, the app must tell this CAS middleware what the
backend base URL is and - based on this base URL, the redirects back to
the app are constructed.

This is husch pfusch implemented and works. Maybe it is necessary to
make it a little nicer. Also I'm not sure if it works without providing
a backendBaseUrl (which would be basically ok if not behind a reverse
proxy)